Wisconsin Diagnostic Breast Cancer Dataset

The WDBC dataset contains quantitative measurements from digitized images of fine needle aspirates (FNA) of breast masses. It is commonly used for classification tasks to distinguish between benign and malignant tumors.

wdbc

Format

A data frame with 569 rows and 32 columns:

radius_mean

Mean of radius

radius_se

Standard error of radius

radius_worst

Worst (largest) radius

texture_mean

Mean of texture

texture_se

Standard error of texture

texture_worst

Worst texture

perimeter_mean

Mean of perimeter

perimeter_se

Standard error of perimeter

perimeter_worst

Worst perimeter

area_mean

Mean area

area_se

Standard error of area

area_worst

Worst area

smoothness_mean

Mean smoothness

smoothness_se

Standard error of smoothness

smoothness_worst

Worst smoothness

compactness_mean

Mean compactness

compactness_se

Standard error of compactness

compactness_worst

Worst compactness

concavity_mean

Mean concavity

concavity_se

Standard error of concavity

concavity_worst

Worst concavity

concave.points_mean

Mean concave points

concave.points_se

Standard error of concave points

concave.points_worst

Worst concave points

symmetry_mean

Mean symmetry

symmetry_se

Standard error of symmetry

symmetry_worst

Worst symmetry

fractal_dimension_mean

Mean fractal dimension

fractal_dimension_se

Standard error of fractal dimension

fractal_dimension_worst

Worst fractal dimension

diagnosis

Factor with levels 'B' and 'M'

Source

Dr. William H. Wolberg, W. Nick Street, and Olvi L. Mangasarian, University of Wisconsin–Madison. Original dataset available at: <https://archive.ics.uci.edu/dataset/17/breast+cancer+wisconsin+diagnostic>